Stripped out conversation manager
